Small test adjustment to account for refactoring
